rename: BridgeGame을 domain 패키지 하위로 옮김
