+++ src/main/java/bridge/domain/BridgeGame.java
@@ -1,5 +1,7 @@
-package bridge;
+package bridge.domain;
 
+import bridge.BridgeMaker;
+import bridge.BridgeRandomNumberGenerator;
 import bridge.domain.GameRound;
 
 import java.util.List;
